backend/requirements.txt
# Flask Web Framework
flask==3.0.0
flask-cors==4.0.0

# Exchange APIs
ccxt>=4.0.0

# Configuration & Security
python-dotenv==1.0.0
cryptography==41.0.7

# Utilities
requests==2.31.0
pyyaml==6.0.1

# Logging
python-json-logger==2.0.7
